TDP-18 moved exceptions to your own namespace
